pyproject.toml
[tool.poetry]
name = "user-management-api"
version = "0.1.0"
description = "REST API for user management with LiteStar"
authors = ["Developer"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1"
litestar = { extras = ["standard"], version = "^2" }
litestar-granian = "^0"
litestar-asyncpg = "^0"
advanced-alchemy = "^0.20"
msgspec = "^0.18.6"
structlog = "^24"
aio-pika = "^9.5.4"
faststream = { extras = ["cli"], version = "^0.5.34" }
asyncpg = "^0.30"
pydantic-settings = "^2"
sniffio = "^1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
